Fixes for clippy
