+++ src/main.rs
@@ -18,10 +18,10 @@ use crate::cartridge::Cartridge;
 use crate::configuration::{Configuration, RuntimeConfig};
 use crate::gpu::color::Color;
 use crate::joypad::JoypadHandler;
-use cpu::cpu::Cpu;
-use gpu::gpu::Gpu;
+use cpu::Cpu;
+use gpu::Gpu;
 use image::ImageBuffer;
-use memory::memory::Memory;
+use memory::Memory;
 use parking_lot::RwLock;
 use piston_window::*;
 use std::sync::{mpsc, Arc};
